[CORE-5043]
