+++ src/app/onboarding/final/final.page.html
@@ -7,14 +7,19 @@
 <ion-content>
   <ion-grid>
     <ion-row>
-      <ion-col size="12" size-md="3">
+      <ion-col size="12" size-md="4">
         <ion-card>
-
+          <ion-card-header>
+            <ion-card-title>Create your account to continue</ion-card-title>
+          </ion-card-header>
+          <ion-card-content>
+            <div id="form"></div>
+          </ion-card-content>
         </ion-card>
 
       </ion-col>
 
-      <ion-col size="12" size-md="9">
+      <ion-col size="12" size-md="8">
         What is next?
       </ion-col>
     </ion-row>
