+++ src/app/overview/overview.component.html
@@ -73,7 +73,6 @@
               {{ tag.name }}<span class="experience-tag-count">{{ tag.count }}</span>
             </button>
           </div>
-          <div id="form"></div>
         </aside>
         <section class="experiences-content col-md-9">
           <div class="experiences-filters">
