+++ src/index.html
@@ -25,7 +25,9 @@
     }
   </script>
   <!-- End of Segment -->
+  <!-- HubSpot -->
   <script charset="utf-8" type="text/javascript" src="//js.hsforms.net/forms/shell.js"></script>
+  <!-- End of HubSpot -->
 </head>
 
 <body>
